move info output to FusionAuthWorkerFactory
